pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinkerrl"
version = "0.1.0"
description = "TinkerRL: A Reproducible Multi-Library Reinforcement Learning Benchmark Suite"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "PES LLM Research Team", email = "arvindcr4@gmail.com"},
]
keywords = [
    "reinforcement-learning",
    "benchmark",
    "reproducibility",
    "deep-learning",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "gymnasium>=0.29.0",
]

[project.optional-dependencies]
trl = [
    "transformers>=4.36.0",
    "trl>=0.7.0",
    "datasets>=2.14.0",
    "peft>=0.5.0",
    "accelerate>=0.24.0",
]
sb3 = [
    "stable-baselines3>=2.1.0",
]
cleanrl = [
    "tensorboard>=2.14.0",
]
tianshou = [
    "tianshou>=0.5.0",
]
d3rlpy = [
    "d3rlpy>=2.0.0",
]
all = [
    "tinkerrl[trl,sb3,cleanrl,tianshou,d3rlpy]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    # Required so `utils.stats` can be imported by the test-core job
    # (utils/stats.py uses pandas / matplotlib / seaborn for the
    # per-experiment analysis tooling).
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
]

[project.urls]
Homepage = "https://github.com/pes-llm-research/tinker-rl-lab"
Documentation = "https://github.com/pes-llm-research/tinker-rl-lab/blob/main/README.md"
Repository = "https://github.com/pes-llm-research/tinker-rl-lab"
Issues = "https://github.com/pes-llm-research/tinker-rl-lab/issues"

[tool.setuptools.packages.find]
# First-party Python packages only. Vendored / integration subtrees
# (atropos, verl, openrlhf, skyrl, trl_integrations, unified,
# agentic-rl-finetuning, ai-scientist-template, research_loop) are kept
# out of the `tinkerrl` wheel on purpose — they ship as source-only
# references from upstream projects and have their own packaging.
include = ["tinkerrl*", "utils*"]
exclude = [
    "atropos*", "verl*", "openrlhf*", "skyrl*", "trl_integrations*",
    "unified*", "agentic-rl-finetuning*", "ai-scientist-template*",
    "research_loop*", "capstone-literature-survey*", "huggingface*",
    "tests*", "paper*", "reports*", "experiments*", "scripts*",
]

[tool.ruff]
line-length = 100
target-version = "py39"
extend-exclude = [
    # Vendored / upstream integrations — not authored by this project.
    "atropos",
    "verl",
    "openrlhf",
    "skyrl",
    "trl_integrations",
    "unified",
    "agentic-rl-finetuning",
    "ai-scientist-template",
    "research_loop",
    "capstone-literature-survey",
    "huggingface",
    # Generated / archived notebooks & result files.
    "*.ipynb",
    "paper/tikz",
    "experiments/collab-results",
    "experiments/modal",
    "experiments/tinker-runs",
    "experiments/10x_structural_ceiling",
    "experiments/notebooks",
    "reports/final",
    "grpo_ablation_results",
]

[tool.ruff.lint]
# Keep the bug-finding checks (F401 unused import, F811 redefinition,
# F821 undefined name) enabled globally so they continue to flag real
# issues in first-party code (``utils/``, ``tests/``, top-level modules).
# Noisy exploratory-code paths (``scripts/**``, ``experiments/**``) scope
# those same rules via ``per-file-ignores`` below; vendored trees are
# excluded entirely via ``extend-exclude`` above.
select = ["F", "W", "I", "UP031"]
ignore = [
    "E501",   # line length — handled by formatter if at all
    "F541",   # f-string without placeholders — used for multi-line format strings
    "F841",   # unused local (research scripts set aside intermediates)
    "W291",   # trailing whitespace (autofixable; CI should not block on it)
    "W293",   # blank line with whitespace
    "I001",   # unsorted imports (let contributors organize imports their way)
]

[tool.ruff.lint.per-file-ignores]
# Scope the noisy notebook/script-style false positives to exploratory code
# paths instead of disabling bug-finding checks for the whole repository.
# First-party code (``utils/``, ``tests/``, library modules) still gets
# F401 / F811 / F821 enforcement.
"scripts/**/*.py"      = ["F401", "F811", "F821"]
"experiments/**/*.py"  = ["F401", "F811", "F821"]
"paper/figures/**/*.py" = ["F401", "F811", "F821"]
# Top-level GRPO research drivers (one-shot scripts, not reusable library
# code; they keep exploratory imports deliberately).
"grpo_*.py"            = ["F401", "F811", "F821"]

[tool.ruff.format]
# Only format what we lint; keep research scripts as the author wrote them
# unless they clearly opt-in via `ruff format`.
exclude = ["experiments/**", "scripts/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
